feat(model): new model
