+++ designer-demo/package.json
@@ -21,6 +21,7 @@
     "@opentiny/vue-renderless": "~3.20.0",
     "@opentiny/vue-theme": "~3.20.0",
     "@vueuse/core": "^9.6.0",
+    "vue-draggable-next": "2.2.1",
     "vue": "^3.4.21"
   },
   "devDependencies": {
