+++ designer-demo/src/configurators/base-form-configurator/BaseFormConfigurator.vue
