+++ designer-demo/src/configurators/common/constants.js
