+++ designer-demo/src/configurators/common/utils.js
