+++ designer-demo/src/configurators/index.js
@@ -1,5 +1,7 @@
 import MyInputConfigurator from './MyInputConfigurator.vue'
+import ModelConfigurator from './model-configurator/ModelConfigurator.vue'
 
 export const configurators = {
-  MyInputConfigurator
+  MyInputConfigurator,
+  ModelConfigurator
 }
