+++ designer-demo/src/configurators/model-api-configurator/ParamsBindGrid.vue
@@ -109,104 +109,6 @@ const rules = computed(() => {
   return rulesMap;
 });
 
-const objectToTreeData = (state, prefix, treeData) => {
-  const tree = treeData;
-  Object.entries(state).forEach(([key, value]) => {
-    if (typeof value === "object" && value !== null && !Array.isArray(value)) {
-      tree.push({
-        label: key,
-        value: {
-          type: "JSExpression",
-          value: `${prefix}.${key}`,
-        },
-        children: [],
-      });
-      objectToTreeData(
-        value,
-        `${prefix}.${key}`,
-        tree[tree.length - 1].children
-      );
-    } else {
-      tree.push({
-        label: key,
-        value: {
-          type: "JSExpression",
-          value: `${prefix}.${key}`,
-        },
-      });
-    }
-  });
-};
-
-const getModelStates = () => {
-  const pageSchema = getPageSchema();
-  const states = pageSchema.state;
-  // 否则当前绑定事件的组件不是模型组件，则可以绑定任意已创建的模型state变量
-  return Object.entries(states)
-    .map(([key, value]) => {
-      if (key.includes("modelState_")) {
-        return {
-          label: key,
-          value: key,
-        };
-      }
-      return null;
-    })
-    .filter((item) => !(item === null));
-};
-
-const typeOptions = ref({
-  data: [
-    {
-      value: "model",
-      label: "模型",
-      children: getModelStates(),
-    },
-    {
-      value: "varible",
-      label: "变量",
-      children: [],
-    },
-  ],
-});
-
-const modelOptions = (modelKey) => {
-  const pageSchema = getPageSchema();
-  const modelState = pageSchema.state[modelKey];
-  const treeData = [];
-  objectToTreeData(modelState, `this.state.${modelKey}`, treeData);
-  return treeData;
-};
-
-const stateOptions = computed(() => {
-  const pageSchema = getPageSchema();
-  const treeData = [];
-  const state = {};
-  for (const key in pageSchema.state) {
-    if (!key.includes("modelState_")) {
-      state[key] = pageSchema.state[key];
-    }
-  }
-  objectToTreeData(state, `this.state`, treeData);
-  return treeData;
-});
-
-const modelTypeChange = (type) => {
-  if (type === "varible") {
-    return { data: stateOptions.value };
-  }
-  if (type && type.includes("modelState_")) {
-    return { data: modelOptions(type) };
-  }
-  return { data: [] };
-};
-
-const varibleChange = () => {
-  tableData.value.forEach((item) => {
-    validateData[item.prop] = item.bindValue ?? "";
-  });
-};
-
 defineExpose({
   validateForm: () => {
     return ruleFormRef.value.validate().then((valid) => {
