+++ designer-demo/src/configurators/model-configurator/MetaListItem.vue
