+++ designer-demo/src/configurators/operator-group-configurator/MetaChildItem.vue
