+++ designer-demo/src/configurators/operator-group-configurator/MetaListItemOp.vue
