+++ designer-demo/src/configurators/operator-group-configurator/MetaListOp.vue
