+++ designer-demo/src/configurators/operator-group-configurator/OperatorGroupConfigurator.vue
