+++ designer-demo/src/configurators/page-configurator/PageConfigurator.vue
